requirements.txt
python-dotenv
streamlit
langchain_community
langchain
pymysql
chromadb
google.generativeai
cryptography